Redux store for word and picture separated
