+++ src/store/index.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware, compose } from "redux";
 
 import ReduxThunk from "redux-thunk";
-import reducer from "./reducer";
+import reducer from "./rootReducer";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
