--- src/store/selectors.js
@@ -1,3 +1,0 @@
-export const selectRandomWord = (state) => state.randomWord;
-export const selectDefinition = (state) =>
-  state.hints ? state.hints.definition : null;
