+++ src/store/word/selectors.js
@@ -1,0 +1,3 @@
+export const selectRandomWord = (state) => state.word.randomWord;
+export const selectDefinition = (state) =>
+  state.word.hints ? state.word.hints.definition : null;
